hello_world/requirements.txt
requests
fastapi
boto3
botocore
python-dotenv